debug_write() method in Config() w/ tests

+++ LazyLetter/configurator.py
@@ -1,5 +1,6 @@
 import os
 import json
+import datetime
 
 from . import utility
 
@@ -12,7 +13,7 @@ class Config(object):
 
     def __init__(self, path_letters='cover letters', path_save='config',
                  greeting="To Whom It May Concern", copy=False, debug=False,
-                 current_filename='default.cfg'
+                 debuglog=None, current_filename='default.cfg'
                  ):
         # designated path to the directory containing the cover letter .txt's
         self.path_letters = self.default_path(path_letters)
@@ -21,9 +22,10 @@ class Config(object):
         self.current_filename = current_filename
         self.greeting = greeting
         self.debug = debug
+        self.debuglog = debuglog
         self.copy = copy
 
-    def default_path(self, path):
+    def default_path(self, path=None):
         """
         Constructs a path relative to the parent of this file based on 2
         default preferences, None or string, or a path input.
@@ -38,6 +40,18 @@ class Config(object):
 
         return result
 
+    def write_debug(self, function_name, message):
+        result = "[DEBUG] " + function_name + ': ' + message
+        if self.debug:
+            print(result)
+        if self.debuglog:
+            filepath = os.path.join(self.default_path(), self.debuglog)
+            with open(filepath, 'a') as f:
+                f.write('['+str(datetime.datetime.now())+'] ' + result)
+                f.close()
+
+        return result
+
     def load_dict(self, indict):
         """
         Loads configuration settings from a dictionary object.
@@ -45,12 +59,16 @@ class Config(object):
         for key in indict:
             if hasattr(self, key):
                 self.__dict__[key] = indict[key]
-            elif self.debug:
-                print("[DEBUG]", __name__, "cannot load invalid key:", key,
-                      "(value:", indict[key], ")",
-                      )
+            else:
+                self.write_debug(self.load_dict.__name__,
+                                 "Cannot load invalid key: "+key+" (value: " +
+                                 indict[key]+")")
 
-    def save(self):
+    def save(self, force=True):
+        """
+        Dumps all attributes in dictionary form to a json text file named
+        with the current_filename string.
+        """
         # check to see if the directories exist
         if not os.path.exists(self.path_save):
             os.makedirs(self.path_save)
@@ -68,13 +86,18 @@ class Config(object):
         f.close()
 
         if os.path.exists(filepath):
-            os.remove(filepath)
+            if force:
+                os.remove(filepath)
+            else:
+                return False
 
         os.rename(temppath, filepath)
 
+        return True
+
     def load(self):
         """
-        Loads a .cfg file into the attributes of the instance, returns T/F
+        Loads a json text file into the attributes of the instance, returns T/F
         depending on file existence.
         """
         filepath = os.path.join(self.path_save, self.current_filename)
@@ -86,13 +109,8 @@ class Config(object):
 
             return True
         except FileNotFoundError as message:
-            if self.debug:
-                print('[DEBUG] Attempted to load',
-                      self.current_filename + ':', message,
-                      )
-            else:
-                print(self.current_filename, "doesn't exist.")
-
+            self.write_debug(self.load.__name__, "Attempted to load " +
+                             self.current_filename+": "+str(message))
             return False
 
     def remove_save(self):
@@ -103,18 +121,23 @@ class Config(object):
         """
         return utility.delete_file(self.path_save, self.current_filename)
 
-    def rename_current_filename(self, new_filename):
+    def rename_current_filename(self, new_filename, force=True):
         """
         Renames the current config's .cfg file to the given filename, switches
         the current_filename to the argument new_filename.
 
-        Returns back the passed argument.
+        Returns back the passed argument on success, otherwise returns the
+        filename that existed prior to this function call.
         """
+        old_filename = self.current_filename
+
         self.remove_save()
         self.current_filename = new_filename
-        self.save()
 
-        return new_filename
+        if self.save(force):
+            return new_filename
+        else:
+            return old_filename
 
     def change_config(self, new_filename):
         """
